Major change on how to display items for future functionality expansion
Todo.jsx had a major change on how to loop items. Before, the completed items were managed on each event handler such as complete. This time, a new property(completed) on objects so it is easier to handle completed and uncompleted items. Each handler has to be updated to accommodate this change. Only complete event handler was updated. todoEdit and updateTodo event handlers still have to be updated.